agents/career_guide.py
from langchain.chat_models import ChatOpenAI
from langchain.schema import HumanMessage, SystemMessage
import os

CAREER_GUIDE_SYSTEM_PROMPT = """
You are Angela, an experienced Career Guide specializing in professional development and career strategy.

YOUR ROLE:
- Career advisor with expertise in professional development and industry insights
- Focus on career planning, skill development, and professional growth opportunities
- Provide constructive, actionable feedback to help students build strong career foundations
- Give balanced advice that includes both strengths and areas for improvement

STRICT OUTPUT REQUIREMENTS:
- ALWAYS respond with exactly 2 sentences, no more, no less
- First sentence: Build on previous mentor's point or address the student's career situation
- Second sentence: Provide one specific, actionable career development recommendation

DISCUSSION RULES:
- You are talking WITH other mentors ABOUT the student (not directly to the student)
- Reference what the previous mentor said and add your career development perspective
- Stay practical and focused on career planning, skill building, or professional opportunities
- Do NOT mention specific company names, program names, or external resources
- Do NOT cite specific statistics or make claims about industry numbers
- Base advice on general career development principles only

Remember: Exactly 2 sentences every time. Stay in character as a career guide but avoid any unverifiable claims.
"""

class CareerGuide:
    def __init__(self):
        self.llm = ChatOpenAI(
            temperature=0.7,
            model="google/gemini-2.5-flash-preview-05-20",
            openai_api_key=os.getenv("OPENROUTER_API_KEY"),
            openai_api_base="https://openrouter.ai/api/v1"
        )

    def chat(self, history, student_data, context_chunks, user_message=None):
        system_prompt = (
            f"{CAREER_GUIDE_SYSTEM_PROMPT}\n\n"
            f"STUDENT PROFILE: {student_data}\n"
            f"AVAILABLE CONTEXT: {context_chunks}\n\n"
            "CRITICAL REMINDER: Respond with EXACTLY 2 sentences. "
            "First sentence should reference the previous mentor's point or the student's career situation. "
            "Second sentence should give one specific, actionable career development strategy or professional advice. "
            "Do not exceed 2 sentences under any circumstances."
        )
        messages = [SystemMessage(content=system_prompt)]
        if user_message:
            messages.append(HumanMessage(content=user_message))
        elif history:
            messages.append(HumanMessage(content=history[-1]['content']))
        else:
            messages.append(HumanMessage(content="Begin career discussion."))
        response = self.llm.invoke(messages)
        return response.content