requirements.txt
# Python dependencies for GVC AI Mentor Roundtable

# Core framework
streamlit>=1.28.0

# Environment management
python-dotenv>=1.0.0

# AI and LLM
openai>=1.0.0
anthropic>=0.3.0

# Vector storage and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Image processing
Pillow>=10.0.0

# HTTP requests
requests>=2.31.0

# File handling
pathlib>=1.0.0

# Additional utilities
typing-extensions>=4.0.0
